Refactoring: args/main, remove most compiler warnings
