+++ src/lib.rs
@@ -1,7 +1,6 @@
 #![feature(fs_walk)]
 #![feature(path_relative_from)]
 #![feature(convert)]
-#![feature(debug_builders)]
 #![feature(plugin)]
 #![feature(path_ext)]
 #![plugin(docopt_macros)]
