+++ src/spag/main.rs
@@ -6,101 +6,21 @@ use std::io::Write;
 use std::collections::hash_map::HashMap;
 use std::path::PathBuf;
 
-use curl::http::handle::Method;
 use curl::http;
-use docopt::Docopt;
+use yaml_rust::Yaml;
 
-use super::request::SpagRequest;
+use super::args;
+use super::args::Args;
 use super::env;
-use super::template;
-use super::request;
 use super::file;
 use super::history;
+use super::request;
+use super::request::SpagRequest;
+use super::template;
 
-use yaml_rust::Yaml;
-use yaml_rust::yaml::Hash;
-
-/// Formats a string which is printed to stderr, and exits with status code 1
-///
-/// ```
-/// error!("Failed to bring it around town: {}", arg);
-/// ```
-macro_rules! error {
-    ($($arg:tt)*) => ({
-        match writeln!(&mut std::io::stderr(), $($arg)*) {
-            Ok(_) => {},
-            Err(x) => panic!("Unable to write to stderr: {}", x),
-        }
-        std::process::exit(1);
-    })
-}
-
-/// Unwraps a Result like try!(), but calls error!("{}", msg) if the result is Err(msg)
-///
-/// ```
-/// let value = try_error!(result);
-/// ```
-macro_rules! try_error {
-    ($expr:expr) => ({
-        match $expr {
-            Ok(val) => val,
-            Err(err) => error!("{}", err),
-        }
-    })
-}
-
-docopt!(Args derive Debug, "
-Usage:
-    spag --help
-    spag env set (<key> <val>)...
-    spag env unset [(<key>)...] [-E]
-    spag env show [<environment>]
-    spag env activate <environment>
-    spag env deactivate
-    spag (get|post|put|patch|delete) <resource> [(-H <header>)...] [-e <endpoint>] [-d <data>] [(--with <key> <val>)...]
-    spag request list [--dir <dir>]
-    spag request show <file>
-    spag request <file> [(-H <header>)...] [-e <endpoint>] [-d <data>] [(--with <key> <val>)...] [--dir <dir>]
-    spag history
-    spag history show <index>
-
-Options:
-    -h, --help                  Show this message
-    -H, --header <header>       Supply a header
-    -e, --endpoint <endpoint>   Supply the endpoint
-    -d, --data <data>           Supply the request body
-    -E, --everything            Unset an entire environment
-    --dir <dir>                 The directory containing request files
-
-Arguments:
-    <endpoint>      The base url of the service, like 'http://localhost:5000'
-    <resource>      The path of an api resource, like '/v2/things'
-    <header>        An http header, like 'Content-type: application/json'
-    <environment>   The name of an environment, like 'default'
-    <index>         An index, starting at zero
-
-Commands:
-    env set         Set a key-value pair in the active environment
-    env unset       Unset one or more keys in the active environment
-    env show        Print out the specified environment
-    env activate    Activate an environment by name
-    env deactivate  Deactivate the environment and return to the default environment
-    get             An HTTP GET request
-    post            An HTTP POST request
-    put             An HTTP PUT request
-    patch           An HTTP PATCH request
-    delete          An HTTP DELETE request
-    request         Make a request using a predefined file
-    request show    Show the specified request file
-    request list    List available request files
-    history         Print a list of previously made requests
-    history show    Print out a previous request by its index
-");
 
 pub fn main() {
-    let args: Args = Args::docopt().decode().unwrap_or_else(|e| e.exit());
-    // println!("{:?}", args);
-
+    let args: Args = args::parse_args();
     if args.cmd_request {
         spag_request(&args);
     } else if args.cmd_history {
@@ -122,7 +42,7 @@ fn spag_env(args: &Args) {
     } else if args.cmd_activate {
         spag_env_activate(&args);
     } else if args.cmd_deactivate {
-        spag_env_deactivate(&args);
+        spag_env_deactivate();
     } else {
         panic!("BUG: Invalid command");
     }
@@ -130,7 +50,7 @@ fn spag_env(args: &Args) {
 
 fn spag_env_set(args: &Args) {
     let use_shortcuts = true;
-    let withs: HashMap<String, String> = get_withs(args);
+    let withs: HashMap<String, String> = args::get_withs(args);
     let withs: HashMap<&str, &str> = withs.iter()
         .map(|(k, v)| (k.as_str(), v.as_str()))
         .collect();
@@ -142,30 +62,30 @@ fn spag_env_set(args: &Args) {
         vals.push(value);
     }
 
-    env::set_in_environment(&args.arg_environment, &args.arg_key, &vals);
-    env::show_environment(&args.arg_environment);
+    try_error!(env::set_in_environment(&args.arg_environment, &args.arg_key, &vals));
+    try_error!(env::show_environment(&args.arg_environment));
 }
 
 fn spag_env_unset(args: &Args) {
-    if args.flag_everything == false {
-        env::unset_in_environment(&args.arg_environment, &args.arg_key);
-        env::show_environment(&args.arg_environment);
+    if !args.flag_everything {
+        try_error!(env::unset_in_environment(&args.arg_environment, &args.arg_key));
+        try_error!(env::show_environment(&args.arg_environment));
     } else {
-        env::unset_all_environment(&args.arg_environment);
-        env::show_environment(&args.arg_environment);
+        try_error!(env::unset_all_environment(&args.arg_environment));
+        try_error!(env::show_environment(&args.arg_environment));
     }
 }
 
 fn spag_env_show(args: &Args) {
-    env::show_environment(&args.arg_environment);
+    try_error!(env::show_environment(&args.arg_environment));
 }
 
 fn spag_env_activate(args: &Args) {
     env::set_active_environment(&args.arg_environment);
-    env::show_environment(&args.arg_environment);
+    try_error!(env::show_environment(&args.arg_environment));
 }
 
-fn spag_env_deactivate(args: &Args) {
+fn spag_env_deactivate() {
     env::deactivate_environment();
 }
 
@@ -198,9 +118,9 @@ fn spag_request(args: &Args) {
 }
 
 fn spag_request_a_file(args: &Args) {
-    let endpoint = try_error!(get_endpoint(args));
-    let dir = try_error!(get_dir(args));
-    let withs: HashMap<String, String> = get_withs(args);
+    let endpoint = try_error!(args::get_endpoint(args));
+    let dir = try_error!(args::get_dir(args));
+    let withs: HashMap<String, String> = args::get_withs(args);
     let withs: HashMap<&str, &str> = withs.iter()
         .map(|(k, v)| (k.as_str(), v.as_str()))
         .collect();
@@ -213,14 +133,14 @@ fn spag_request_a_file(args: &Args) {
 
     match file::load_yaml_string(&yaml_string) {
         Ok(y) => {
-            let method = get_string_from_yaml(&y, &["method"]);
-            let uri = get_string_from_yaml(&y, &["uri"]);
+            let method = args::get_string_from_yaml(&y, &["method"]);
+            let uri = args::get_string_from_yaml(&y, &["uri"]);
 
             // the request body can be overridden by the --data flag.
             //
             // todo? because docopt defaults to an empty string if the data flag isn't given,
             // we can't tell if the user is trying to override the body to be empty.
-            let data = try_error!(get_data(args));
+            let data = try_error!(args::get_data(args));
             let body =
                 if !data.is_empty() {
                     data
@@ -232,7 +152,7 @@ fn spag_request_a_file(args: &Args) {
                     }
                 };
 
-            let headers = try_error!(resolve_headers(args, &y));
+            let headers = try_error!(args::resolve_headers(args, &y));
 
             let mut req = SpagRequest::new(request::method_from_str(&method), endpoint, uri);
             try_error!(req.add_headers(headers.iter()));
@@ -244,7 +164,7 @@ fn spag_request_a_file(args: &Args) {
 }
 
 fn spag_request_show(args: &Args) {
-    let dir = try_error!(get_dir(args));
+    let dir = try_error!(args::get_dir(args));
     let filename = try_error!(request::get_request_filename(&args.arg_file, &dir));
     // TODO: read_file() panics. don't do that.
     let contents = file::read_file(&filename);
@@ -252,7 +172,7 @@ fn spag_request_show(args: &Args) {
 }
 
 fn spag_request_list(args: &Args) {
-    let dir = try_error!(get_dir(args));
+    let dir = try_error!(args::get_dir(args));
     let filenames = try_error!(file::walk_dir(&dir));
     let mut yaml_files: Vec<&PathBuf> = filenames.iter()
         .filter(|p| p.to_str().unwrap().ends_with(".yml"))
@@ -269,19 +189,19 @@ fn spag_request_list(args: &Args) {
 fn spag_method(args: &Args) {
     // untemplate the resource
     let use_shortcuts = true;
-    let withs: HashMap<String, String> = get_withs(args);
+    let withs: HashMap<String, String> = args::get_withs(args);
     let withs: HashMap<&str, &str> = withs.iter()
         .map(|(k, v)| (k.as_str(), v.as_str()))
         .collect();
     let resource = try_error!(template::untemplate(&args.arg_resource, &withs, use_shortcuts));
 
-    let method = get_method_from_args(args);
-    let endpoint = try_error!(get_endpoint(args));
+    let method = args::get_method_from_args(args);
+    let endpoint = try_error!(args::get_endpoint(args));
     let mut req = SpagRequest::new(method, endpoint, resource);
-    let headers = try_error!(resolve_headers_no_request_file(args));
+    let headers = try_error!(args::resolve_headers_no_request_file(args));
     try_error!(req.add_headers(headers.iter()));
 
-    let body = try_error!(get_data(args));
+    let body = try_error!(args::get_data(args));
     req.set_body(body);
     do_request(&req);
 }
@@ -290,144 +210,6 @@ fn do_request(req: &SpagRequest) {
     let mut handle = http::handle();
     let resp = req.prepare(&mut handle).exec().unwrap();
     println!("{}", String::from_utf8(resp.get_body().to_vec()).unwrap());
-    history::append(req, resp);
+    try_error!(history::append(req, resp));
 }
 
-fn get_method_from_args(args: &Args) -> Method {
-    if args.cmd_get { Method::Get }
-    else if args.cmd_post { Method::Post }
-    else if args.cmd_put { Method::Put }
-    else if args.cmd_patch { Method::Patch }
-    else if args.cmd_delete { Method::Delete }
-    else { panic!("BUG: method not recognized"); }
-}
-
-fn get_endpoint(args: &Args) -> Result<String, String> {
-    // passing -e ENDPOINT overrides everything else
-    if !args.flag_endpoint.is_empty() {
-        Ok(args.flag_endpoint.to_string())
-    } else {
-        let env = try!(env::load_environment(""));
-
-        if let Some(e) = env["endpoint"].as_str() {
-            Ok(e.to_string())
-        } else {
-            Err("Endpoint not set".to_string())
-        }
-    }
-}
-
-fn get_dir(args: &Args) -> Result<String, String> {
-    // passing in --dir overrides everything else
-    if !args.flag_dir.is_empty() {
-        Ok(args.flag_dir.to_string())
-    } else {
-        let env = try!(env::load_environment(""));
-
-        if let Some(e) = env["dir"].as_str() {
-            Ok(e.to_string())
-        } else {
-            Err("Request directory not set".to_string())
-        }
-    }
-}
-
-fn get_data(args: &Args) -> Result<String, String> {
-    let use_shortcuts = true;
-    let withs: HashMap<String, String> = get_withs(args);
-    let withs: HashMap<&str, &str> = withs.iter()
-        .map(|(k, v)| (k.as_str(), v.as_str()))
-        .collect();
-    Ok(try!(template::untemplate(&args.flag_data, &withs, use_shortcuts)))
-}
-
-fn get_headers_from_request(request_yaml: &Yaml) -> Result<HashMap<String, String>, String> {
-    let default_hash = &Yaml::Hash(Hash::new());
-    let mut result: HashMap<String, String> = HashMap::new();
-    let request_file_headers = env::get_nested_value(&request_yaml, &["headers"]).unwrap_or(default_hash);
-    if let &Yaml::Hash(ref h) = request_file_headers {
-        for (k, v) in h.iter() {
-            if let (&Yaml::String(ref key), &Yaml::String(ref value)) = (k, v) {
-                result.insert(key.to_string(), value.to_string());
-            }
-        }
-    }
-    Ok(result)
-}
-
-fn get_headers_from_environment() -> Result<HashMap<String, String>, String> {
-    let default_hash = &Yaml::Hash(Hash::new());
-    let mut result: HashMap<String, String> = HashMap::new();
-    // TODO: case insensitivity
-    // be sure not to fail if we fail to load the env.
-    let env = env::load_environment("").unwrap_or(Yaml::Hash(Hash::new()));
-    let env_headers = env::get_nested_value(&env, &["headers"]).unwrap_or(default_hash);
-    if let &Yaml::Hash(ref h) = env_headers {
-        for (k, v) in h.iter() {
-            if let (&Yaml::String(ref key), &Yaml::String(ref value)) = (k, v) {
-                result.insert(key.to_string(), value.to_string());
-            }
-        }
-    }
-    Ok(result)
-}
-
-fn get_headers_from_args(args: &Args) -> Result<HashMap<String, String>, String> {
-    let use_shortcuts = true;
-    let mut result: HashMap<String, String> = HashMap::new();
-    let arg_headers: Vec<(&str, &str)> =
-        try_error!(args.flag_header.iter().map(|s| request::split_header(s)).collect());
-    for &(k, v) in arg_headers.iter() {
-        let v = try_error!(template::untemplate(&v, &HashMap::new(), use_shortcuts));
-        result.insert(k.to_string(), v.to_string());
-    }
-    Ok(result)
-}
-
-/// Build a single list of headers from the environment, the request yaml, and arguments.
-fn resolve_headers(args: &Args, request_yaml: &Yaml) -> Result<Vec<String>, String> {
-    let request_headers = try!(get_headers_from_request(request_yaml));
-    let env_headers = try!(get_headers_from_environment());
-    let arg_headers = try!(get_headers_from_args(args));
-    let mut result: HashMap<String, String> = HashMap::new();
-    // start with headers in the environment
-    result.extend(env_headers);
-    // headers in the request override headers in the environment
-    result.extend(request_headers);
-    // headers in arguments override everything
-    result.extend(arg_headers);
-
-    // format headers as "<key>: <val>"
-    let str_headers: Vec<String> = result.iter()
-        .map(|(k, v)| format!("{}: {}", k, v))
-        .collect();
-    Ok(str_headers)
-}
-
-fn resolve_headers_no_request_file(args: &Args) -> Result<Vec<String>, String> {
-    resolve_headers(args, &Yaml::Hash(Hash::new()))
-}
-
-fn get_string_from_yaml(y: &Yaml, keys: &[&str]) -> String {
-    match env::get_nested_value(y, keys) {
-        Some(&Yaml::String(ref m)) => { m.to_string() },
-        Some(ref s) => {
-            error!("Invalid value '{:?}' for key {:?} in request file", s, keys);
-        },
-        _ => {
-            error!("Missing key {:?} in request file", keys);
-        },
-    }
-}
-
-fn get_withs(args: &Args) -> HashMap<String, String> {
-//    println!("key={:?} val={:?}", args.arg_key, args.arg_val);
-    let use_shortcuts = true;
-    let mut withs = HashMap::new();
-    for (k, v) in args.arg_key.iter().zip(args.arg_val.iter()) {
-        let v = try_error!(template::untemplate(&v, &HashMap::new(), use_shortcuts));
-        withs.insert(k.to_string(), v.to_string());
-    }
-//    println!("{:?}", withs);
-    withs
-}
